Fix #25: Ensure table file header definition is consistent
To ensure consistency in the size of the CFE_TBL_FileHdr_t struct,
this should be constrained to use only mission-scope definitions.

Also simplifies the structure definitions by _NOT_ padding strings
out to 32 bit multiples.  The default size of the strings are
already 32 bit multiples so this is just unnecessary complexity.
There is also no major issue if not 32 bit aligned, as the
compiler will add it automatically where needed.

Add a note in the related cfe_mission_cfg.h descriptions to
affected values, that these should be kept as a multiple of 4
to maintain alignment.

Updated for 2019-09-16 code review:
 - Added compile-time verification that the configuration
   directives are actually multiples of 4.