Fix #24: Add strict warning flags
Add extra compile options for mission scope and arch scope.

These are separated to support cross compile environments that
do not/cannot use the same flags on both builds.

For "mission" build the targets are never cross compiled, only
built for the native host machine.  It should be safe to assume
a compiler in the GCC family and the strict warnings should
_always_ be applicable here.

For "arch" build the options are compiler vendor dependent.  The
file as-supplied can only be used if all the target cross compilers
are in the same family and support the same warning options.
However, this file can be modified without affecting the options
used for the host side tools.