Fix #581, Propagate the OSAL compile definitions to CFE build
Use the INTERFACE_COMPILE_DEFINITIONS and INTERFACE_INCLUDE_DIRECTORIES
properties from the osal target and apply them to the entire CFE build.

At this time, the OSAL library build does not use/export these properties
so this is effectively a no-op for the CFE build and can be merged with
no effect.  However, in a future version, the OSAL library will export
these interface properties and this will become important.