Fix #245, Opaque CFE_SB_MsgId_t values
Treat the CFE_SB_MsgId_t type as an opaque/abstract value and do not
assume it is an integer.

This change offers two modes of operation, where CFE_SB_MsgId_t is
defined as a simple integer and is backward compatible, or defined
as a type safe structure.  In type safe mode, passing an integer to
an API requiring a CFE_SB_MsgId_t value will result in an error.
The macros and conversion functions can be used with either mode,
allowing a transition for applications.