Fix #456, Perf log threading and concurrency issues
Replaces the `OS_IntLock` with a standard OSAL mutex for protecting
the shared/global perflog data structure.  This may introduce unexpected
task switches when contention occurs, but it ensures proper exclusion
with respect to the data structures.

Removes the temporary child worker task that was spawned for writing
the log data to a file, and replace with a more generic CFE ES
background task.  The background task is started at boot and
pends on a semaphore until there is work to do.

The background performance log dump is implemented as a state machine
which is called repeatedly over time from the background job task.  This
performs a limited amount of work on each invocation, and resumes where
it left from the previous invocation.