Fix #746, simplify dependency handling
Put more dependencies into the "mission_defaults.cmake" file for
more visibility and ease of configuration.  This now includes all
"implicit" modules such as cfe-core, osal, and psp.

Also push the calls to "generate_config_includefile" to a sub-script
which can be distributed with each app and evaluated as part of the
build.  This reduces dependencies on special naming conventions like
"fsw/mission_inc" and "fsw/platform_inc", and apps can explicitly manage
the files that users are expected to override.