Fix #778, Add CFE_Assert library and test module
Provides ability to load UT assert as a CFE app, and an example
of using this to test some basic CFE ES functions.

This introduces a separate library for the basic UT assert functions
(cfe_assert) and an app that executes the test (cfe_testrunner) rather
than combining these into a single module.