Fix #894, add resource ID type
Add a new typedef "CFE_ES_ResourceID_t" that can
replace uint32 for all ID storage and manipulation.
Initially this is just an alias to uint32 for
backward compatibility.

Convert all APIs that accept an ID to use the new
typedef.

This also reserves the value "0" as an undefined ID,
and gives a separate base value for each resource type.
Therefore even though the type is still uint32, the
different resource IDs can still be distingushed.