Fix #906, Update variable checks in read_targetconfig
This was using "DEFINED" to check if these variables were set.
Problem discovered is that this is always true because "SIMULATION"
is a cache var set from an environment variable, so it ALWAYS defined,
it is just empty if not being used.

Fix is to use if (SIMULATION) rather than if (DEFINED SIMULATION) which
should only be true if the string is not empty, as intended.