Fix #965, remove OSAL ID from App/LibInfo struct
The "osal_id_t" type isn't defined in any of the CFE message/interface
header files for use within telemetry.

This ID is an ephemeral runtime value and is not relevant to a ground
system or anything outside CFE, so it makes most sense to simply
remove this ID from the telemtry structure.  Note that all commands
are name-based, not ID-based, hence why this ID is not really useful.