Fix #1215, remove task registration calls
Explicit task registration is no longer necessary, since all
required actions can be done before invoking the entry point.

This moves the invocation of CFE_PSP_AttachExceptions() from
the registration function to the pre-entry function, this was
the only remaining action in task registration.

All references to task registration in code, docs, and tests
are removed.