Fix #1281, functional test as app instead of library
Merge the "testrunner" app and put the runner logic into cfe_assert lib.
Converts the "testcase" library into an app that simply calls the logic
in cfe_assert to do the test.

This means that the test can be run on demand by issuing the ES
start app command, rather than putting it into the startup script.