Fix #1301, remove option for "osal_compatible"
Currently it is assumed/required that resource IDs follow the "osal compatible"
pattern.  Perhaps in a future version this could change, but for now it is
simpler to just require this configuration, rather than having an option
with only one choice.

This renames/moves the header file but the content is not fundamentally changed.