Fix #1317, memory pool pointer type
Changes the type of pointer in the API from uint8* to void* to be
more consistent and easier to use.  Should be backward compatible.

This also updates the doxygen documentation for this parameter, as
it was specifying a 32-bit alignment requirement whereas the
alignment requirement is processor dependent - 64 bit processors
typically will need 64 bit alignment.  Link to the macro which
is intended to aid in aligning the static pool buffer.