Fix #1345, exception logic when app/task is not found
Adds comments to describe the logic when an exception cannot
be traced back to a specific app, in that it should fall back
to the PSP reset.

Restarting only an app is a special opt-in case that is only
done if specifically selected when starting the app, and the
exception is also positively traced back to that app.