Fix #1436, table object file name
This uses a sub-script to call elf2cfetbl which avoid relying
on a shell glob.  More complicated but should be more robust if/when
someone renames a table.

Note: A proper/simpler fix would still be to use TARGET_OBJECTS generator
expression, but that can't happen until Ubuntu 18.04 (w/ ver 3.10) or older
versions are no longer supported.