Fix #1343, OSAL config file simplification
The CFE framework build uses mostly default values for OSAL compile
time configuration.  There are just a few exceptions where the
CFE build is tuned for a more debug-friendly result.

This updates the configuration files in "sample_defs" to reflect
only those values which are set to a non-default value, along with
information about why the configurable item is changed. Currently,
these are only items related to debugging.

This change also includes more documentation/comments about how this
configuration facility works and where to find information on the
complete set of options available (which is OSAL version dependent).