Fix #1417, generated coverage stubs for CFE core
Update CFE core stub libraries to use generated stubs, using the generator
script now part of UT Assert.  All existing stub logic is converted to
a default handler routine.

This also entailed moving all of the internal API prototypes to be
in the "core_private" interface lib rather than in "core_api".  The
only reason this was not done earlier was due to issues with the stubs,
but that is alleviated when using generated stubs.