Fix #1538, add capability to generate multiple tables
Adds a "install_custom.cmake" hook that can be put into a CPU-specific
subdirectory under the "defs" directory, that can perform extra
installation steps as required for the project/mission.

Tweaks the "add_cfe_tables" function such that it can also be called from
the install_custom.cmake script to generate additional/alternative table
binary files for that CPU.

The main update is that it uses the "APP_NAME" parameter to associate
the table files with the app library, so the same set of include files
can be used.  This relies on the target-scope properties being used.
Historically that string wasn't verified, it could have been any unique
string, but now it should match the app if this is to work as expected.