Fix #1589, Separate warning file for doc builds
Use a unique file name for the doxygen warnings, instead
of "warnings.log", so the mutiple documentation targets
will not overwrite the other log file.

Use a prefixed name so that if the log file is uploaded
to the build artifacts, it will be identifiable.

Detail Design warnings in: detaildesign-warnings.log
Users Guide warnings in: cfe-usersguide-warnings.log