Fix #471, improve SB coverage test
Add test cases to exercise all functions, lines, and branches to
the extent reasonably possible.  Improves the coverage stats
significantly:

  functions 98.9% -> 100%
  lines 96.4% -> 99.8%
  branches 87.1% -> 94.9%

Remaining uncovered lines/branches are not possible to be reached
due to the way the code is structured, or because it would require
an alternate implementation of SBR (note that SB+SBR are currently
tested as a single unit, even though they are technically separate
modules now).  For example, the "direct" SBR implementation cannot
have collisions, hence the collision handling in SB cannot be
reached.  Making stubs for SBR may allow this to be tested.