Fix #1606, update documentation for CFE_ES_GetPoolBufInfo
Corrects the return type documentation, on success this function
returns the size of the buffer, it does not return CFE_SUCCESS.

Additionally, this updates the general description of the pool
implementation to reflect the variances in platform architectures
(i.e. the buffer descriptor/overhead may be 12 bytes on a CPU
with 32 bit size_t, but will be greater on a CPU with a 64 bit
size_t).