Fix #1396, scrub command documentation
Update documentation for all command codes in CFE core apps, in
particular the error conditions, to confirm it matches the
actual implementation.

For common operations that apply to all commmands (length verification)
this is now put into the users guide as a general statement rather
than repeating the info in every command.