Fix #1812, Add SB API test cases
Adds SB functional test cases to cover all missing items that
were identified as part of the scrub in issue #1724.

Where a specific condition is not testable because it requires
a failure of another subsystem, it is marked as `covtest` to
indicate it is only verifiable in coverage test environment.