Fix #1829, confirm CFE_SB_PEND_FOREVER flag
Use the CFE_SB_PEND_FOREVER flag on a few of the calls to CFE_SB_ReceiveBuffer
to confirm that this works as expected.  Previously only CFE_SB_POLL and
a nonzero timeout were used.

Note this can only be used when it is known/expected that the message
queue is not empty.  If there is any possiblility that the queue is empty,
then this cannot be used or else the test may block.  This also means that
if the test fails, it may fail by blocking and never finishing the test
rather than an actual FAIL testcase.