Fix #1949, update msgid testcase to match implementation
Updates the test case for "CFE_MSG_SetMsgId()" to write a value that
is outside the set of storable values for MsgId.  This test has a
caveat that it is somewhat implementation-dependent, but by passing
a MsgId value with all bits set, at least one of those bits is
likely not correlated with a real header bit.