Fix #2228, Use correct message types
Update remaining cases where the CFE_MSG_CommandHeader_t was used for a
command that had no arguments, replacing it with a real data type that
is named according to the command.

In particular, this introduces a "SendHkCmd" type into all modules,
where previously the HK command used CFE_MSG_CommandHeader_t directly.