Fix #2240, improve 64-bit memory address handling in CMD/TLM
The "CFE_ES_MemAddress_t" and "CFE_ES_MemOffset_t" types were intended
to provide a path for easily upgrading the CMD/TLM structs from 32-bit
to 64-bit memory addresses.  However, this type was a bit overused and
in some of those use-cases (e.g. in TBL header) it assumed that the type
was 32-bits during the byte swap ops.  As a result, the type could not
be changed to 64 bits as intended.

This reverts those cases in TBL back to uint32 (meaning that tables will
still be limited to 32 bit sizes, even on 64 bit CPUs) but otherwise the
addresses and sizes in ES/SB telemetry can grow to 64 bits as intended.

For unit tests, correct operation depends on the availablily of an updated
test macro that can compare integers as "size_t" type. (as opposed to
uint32).