Fix #2284, improve add_cfe_tables function
Improves the functionality and adds additional documentation about
how this function is intended to work.  The improvements add some
flexibility and intelligence and should be backward compatible with
existing use cases.

The add_cfe_app() function now adds an additional interface target that
can be referenced when building tables, and this includes the
directory-scope properties from the original app build.  Therefore,
calls to add_cfe_tables from other (non-app) contexts can get the full
set of include directories.

This also makes the target name simpler when adding custom properties,
it is simply "${APP_NAME}.table"

Finally, instead of invoking the table tool (elf2cfetbl) directly in the
context of the table build rule, it generates a custom makefile rule
instead, which is called from the top-level (mission) scope to do the
conversions.