Fix #2289, implement common search routine for config files
Adds a 'cfe_locate_implementation_file()' function that provides a
consistent means to find a file within the ${MISSION_DEFS} subdirectory.

Similar locate/search logic was necessary in a few places, for
generate_config_includefile as well as add_cfe_tables, but these were
not consistent in the paths or ordering used.  Using this function
should make them consistent.