Fix #2423, avoid use of abspath make function
The abspath function does not work in GNU make v3.80, which is still
used as it is packaged with the Wind River tools as part of VxWorks 6.x.

The $(abspath ...) function was used in two places related to generating
tables - first for extraction of the ELF file from an intermediate
lib, second for conversion of that ELF file to a table file.  In the
first case, the path should already be absolute, as it comes from CMake.
In the second case, the path should always be relative, because it is
intended to match another pattern rule starting with "elf/%".  For the
second rule, prefixing with $(CURDIR) should achieve the same effect.