Fix #2507, add EDS cmake hooks
Adds CFE_EDS_ENABLED_BUILD option at the top level, along with
other logic that is tied in only when this is set TRUE.

By default it is set to FALSE, so it does not change any existing
workflows yet.

Also added are the keys for two configuration values for reflecting
EDS DB objects (pointers) and some minor cleanup.