Fix #2514, change CFE_MSG_Message from union to struct
Having this abstract type as a "struct" makes it match the Command
and Telemetry abstract types.  Futhermore, it better conveys the
intent that this is an abstract object and should not be directly
used or accessed in other ways.

It may still be implemented as a union underneath (depending on
how MSG module chooses to implement) but that is hidden from
public API.  In the case of the default MSG module implementation,
there are just a handful of cases where it is accessed internally
as bytes, and those are simple enough to do with a cast.