Fix #2516, propagate stack pointer for child tasks
Update CFE_ES_CreateChildTask to propagate the user-supplied stack
pointer to the underlying OS_TaskCreate call.  Also adds a functional
test to check that the memory address of a local variable within a
child task resides within the expected stack buffer.

NOTE: this requires an additional fix to POSIX OSAL to make it work
on that platform.