Fix #68, Implement transaction event reporting API
Implement deferred event reporting support in the table services
transaction object.  This can be used when off-nominal sitations
are encountered, which should be reported in an event, but the
global data is not in a state where external modules should be called.
For example, if the global registry is locked, then extenal calls
outside of TBL services should not be invoked until it is unlocked.

This just includes the API and unit tests.  It does not change existing
code to utilize the feature yet.