Fix #2651, initialize pipename buffer
When printing events that have no context info (for example, if it
failed one of the very early sanity checks, and never got to the point
of even starting the action), the "PipeName" buffer in the event report
was not initialized.  This initializes the buffer to [n/a] in this case
so it does not attempt to print random data in the event text.