Fix #25, update unit tests
When testing this change along with other upcoming changes (EDS-related)
a number of unit testing issues appeared.  This commit includes a bunch
of unit test updates for TBL to avoid those issues in the future.

There is only one small change to FSW, which was revealed due to the
improvement in unit test cases.  The "GetInfo" call was not setting
the update time nor CRC output values if the table was not loaded.  This
includes a fix to set those values to a proper value, not just leave
them unset.

All other changes are strictly in the unit test domain, improving the
test coverage and maintainability of the tests.  The only other item of
note was that this un-does a hack that was in the msg module, as the
ability to pass multiple test source files to the function is now
properly included.