Fix #47, remove LoadPending separate flag
As a separate flag, it can become inconsistent with the actual state of
the record.  This converts it to a derived value by looking at the state
of the underlying record, rather than a separate flag.  Thus, there is
now a single source of truth - the state of the record itself.

Note that "LoadPending" is a sub-state of "LoadInProgress".
LoadInProgress is true any time there is a load buffer, but LoadPending
is true only when that buffer is ready to be activated (i.e. passed all
validations and any other checks).

The function is now named "PendingActivation" to make this distinction
more clear - as LoadPending and LoadInProgress are not really
distinctive.