Fix #143, add ElementSize to filedef struct and macro
This adds another member to the FileDef struct (used for building
tables) that indicates the size of each individual element in the table.
This is useful for discerning if the object is an array of records or
just a single object.

This is an added member and does not change any existing members so it
is fully backward compatible.