Fix #156, Add logic to handle numbered tables
Allow a table name with a numeric suffix.  If there is no exact match
then check for a numeric value and lookup the same name without the
number.  This is needed for apps that have multiple instances of a table
that all have the same basic definition, for example the MD DWELL_TABLE
which has variable number of instances.