Fix #158, update FS header fields to force big endian
The CFE FS header is always expected to be in network byte order
regardless of the host CPU arch.

As the eventual goal is to allow using EDS-generated data structs (such
as tables) to be used with non-EDS builds, the approach here is to make
these fields always big endian so it matches the non-EDS implementation
of CFE_FS_ReadHeader().