Fix #393, change default memory address size to 64 bits
This also moves the definition of these two types and the associated
macros so that the user does not need to override the entire
extern_typedefs.h header if they want to stay with 32 bit addressing.

To change the type used for the MemAddress, one only needs to override
the "cfe_es_memaddress.h" file and provide the typedef and macro.

This also provides a fallback to the old 32 bit sizes, for users where
the 64 bit size is problematic.