Fix #153, report actual EDS status for component
Adds a lookup of the component name to determine if there is an EDS
package of the same name in the catalog.

If there is a match, then EDS will be functional on this app (active)

If there is no match, then it reports as inactive.  This will be the
case if EDS is globally enabled but some components do not have EDS
files associated with them.