Fix #174, add API to externalize table handles
The CFE_TBL_HandleId_t is the type that should be used in all interfaces
going forward (e.g. messages, tables).  The CFE_TBL_Handle_t is for API
compatibility - and may be a typedef to the same thing.

For cases where it is not the same thing, this adds an API to convert
the CFE_TBL_Handle_t value into a CFE_TBL_HandleId_t value for putting
into messages or tables, and vice versa for getting from a message or
table.